@@ -19,14 +19,14 @@ namespace OPS.API.Controllers
     {
         private IWebHostEnvironment Environment { get; }
         public IConfiguration Configuration { get; }
-        public FornecedorDao dao { get; }
+        public FornecedorRepository dao { get; }
 
         public FornecedorController(IConfiguration configuration, IWebHostEnvironment env)
         {
             Environment = env;
             Configuration = configuration;
 
-            dao = new FornecedorDao();
+            dao = new FornecedorRepository();
         }
 
         [HttpGet]
@@ -86,70 +86,70 @@ namespace OPS.API.Controllers
         private const string paginaCaptcha = "captcha/gerarCaptcha.asp";
         //private const string paginaQuadroSocietario = "Cnpjreva_qsa.asp";
 
-        [HttpGet]
-        //[IgnoreCacheOutput]
-        [Route("Captcha/{value}")]
-        public string Captcha(string value)
-        {
-            CookieContainer _cookies = new CookieContainer();
-            var htmlResult = string.Empty;
+        //[HttpGet]
+        ////[IgnoreCacheOutput]
+        //[Route("Captcha/{value}")]
+        //public string Captcha(string value)
+        //{
+        //    CookieContainer _cookies = new CookieContainer();
+        //    var htmlResult = string.Empty;
 
-            using (var wc = new CookieAwareWebClient(_cookies))
-            {
-                wc.Headers[HttpRequestHeader.UserAgent] = "Mozilla/4.0 (compatible; Synapse)";
-                wc.Headers[HttpRequestHeader.KeepAlive] = "300";
-                htmlResult = wc.DownloadString(urlBaseReceitaFederal + paginaPrincipal);
-            }
+        //    using (var wc = new CookieAwareWebClient(_cookies))
+        //    {
+        //        wc.Headers[HttpRequestHeader.UserAgent] = "Mozilla/4.0 (compatible; Synapse)";
+        //        wc.Headers[HttpRequestHeader.KeepAlive] = "300";
+        //        htmlResult = wc.DownloadString(urlBaseReceitaFederal + paginaPrincipal);
+        //    }
 
-            if (htmlResult.Length > 0)
-            {
-                using (var wc2 = new CookieAwareWebClient(_cookies))
-                {
-                    wc2.Headers[HttpRequestHeader.UserAgent] = "Mozilla/4.0 (compatible; Synapse)";
-                    wc2.Headers[HttpRequestHeader.KeepAlive] = "300";
-                    byte[] data = wc2.DownloadData(urlBaseReceitaFederal + paginaCaptcha);
+        //    if (htmlResult.Length > 0)
+        //    {
+        //        using (var wc2 = new CookieAwareWebClient(_cookies))
+        //        {
+        //            wc2.Headers[HttpRequestHeader.UserAgent] = "Mozilla/4.0 (compatible; Synapse)";
+        //            wc2.Headers[HttpRequestHeader.KeepAlive] = "300";
+        //            byte[] data = wc2.DownloadData(urlBaseReceitaFederal + paginaCaptcha);
 
-                    CacheHelper.Add("CookieReceitaFederal_" + value, _cookies);
+        //            CacheHelper.Add("CookieReceitaFederal_" + value, _cookies);
 
-                    return "data:image/jpeg;base64," + Convert.ToBase64String(data, 0, data.Length);
-                }
-            }
+        //            return "data:image/jpeg;base64," + Convert.ToBase64String(data, 0, data.Length);
+        //        }
+        //    }
 
-            return string.Empty;
-        }
+        //    return string.Empty;
+        //}
 
-        [HttpPost]
-        //[IgnoreCacheOutput, InvalidateCacheOutput(nameof(Consulta))]
-        [Route("ConsultarDadosCnpj")]
-        public dynamic ConsultarDadosCnpj(JsonDocument jsonData)
-        {
-            if (jsonData == null) throw new ArgumentNullException(nameof(jsonData));
+        //[HttpPost]
+        ////[IgnoreCacheOutput, InvalidateCacheOutput(nameof(Consulta))]
+        //[Route("ConsultarDadosCnpj")]
+        //public dynamic ConsultarDadosCnpj(JsonDocument jsonData)
+        //{
+        //    if (jsonData == null) throw new ArgumentNullException(nameof(jsonData));
 
-            JsonElement root = jsonData.RootElement;
+        //    JsonElement root = jsonData.RootElement;
 
-            string aCNPJ = root.GetProperty("cnpj").ToString();
-            CookieContainer _cookies = CacheHelper.Get<CookieContainer>("CookieReceitaFederal_" + aCNPJ);
+        //    string aCNPJ = root.GetProperty("cnpj").ToString();
+        //    CookieContainer _cookies = CacheHelper.Get<CookieContainer>("CookieReceitaFederal_" + aCNPJ);
 
-            var oFormatarDados = new FormatarDados();
-            var fornecedor = oFormatarDados.ObterDados(_cookies, aCNPJ, root.GetProperty("captcha").ToString());
+        //    var oFormatarDados = new FormatarDados();
+        //    var fornecedor = oFormatarDados.ObterDados(_cookies, aCNPJ, root.GetProperty("captcha").ToString());
 
-            // Testar
-            ////now get cache instance
-            //var cache = Configuration.CacheOutputConfiguration().GetCacheOutputProvider(Request);
-            ////and invalidate cache for method "Get" of "FornecedorController"
-            //cache.RemoveStartsWith(Configuration.CacheOutputConfiguration().MakeBaseCachekey("FornecedorController", "Get"));
+        //    // Testar
+        //    ////now get cache instance
+        //    //var cache = Configuration.CacheOutputConfiguration().GetCacheOutputProvider(Request);
+        //    ////and invalidate cache for method "Get" of "FornecedorController"
+        //    //cache.RemoveStartsWith(Configuration.CacheOutputConfiguration().MakeBaseCachekey("FornecedorController", "Get"));
 
-            if (fornecedor != null) //encontrou?
-            {
-                var _fornecedor = dao.Consulta(fornecedor.id);
-                var _quadro_societario = dao.QuadroSocietario(fornecedor.id);
+        //    if (fornecedor != null) //encontrou?
+        //    {
+        //        var _fornecedor = dao.Consulta(fornecedor.id);
+        //        var _quadro_societario = dao.QuadroSocietario(fornecedor.id);
 
-                return new { fornecedor = _fornecedor, quadro_societario = _quadro_societario };
-            }
-            else
-            {
-                return null;
-            }
-        }
+        //        return new { fornecedor = _fornecedor, quadro_societario = _quadro_societario };
+        //    }
+        //    else
+        //    {
+        //        return null;
+        //    }
+        //}
     }
 }
