@@ -2,7 +2,9 @@
   <div class="container" id="sobre-page">
     <br />
     <div id="info-geral">
-      <h3 class="media-heading page-title">Conheça a Operação Política Supervisionada</h3>
+      <h3 class="media-heading page-title">
+        Conheça a Operação Política Supervisionada
+      </h3>
       <br />
       <div class="row">
         <div class="col-md-6">
@@ -38,108 +40,79 @@
             <div class="card-body">
               <p class="text-justify style2">
                 A
-                <strong>Operação Política Supervisionada</strong> fiscaliza de forma detalhada os gastos realizados via CEAP ou CEAPS. Até o momento já foram economizados quase R$ 6 milhões do dinheiro público graças a estas fiscalizações e às exigências feitas diretamente aos parlamentares para que devolvam o dinheiro público indevidamente utilizado.
+                <strong>Operação Política Supervisionada</strong> fiscaliza
+                detalhadamente gastos realizados com verba pública por deputados
+                federais, senadores e de deputados estaduais de alguns estados.
               </p>
-              <p
-                class="text-justify style2"
-              >A OPS conta com a ajuda de seus colaboradores, espalhados pelo Brasil, para o levantamento de informações necessárias para a conclusão de fiscalizações, como por exemplo, o envio de fotos de endereços suspeitos em diversas cidades do país. Além disso, qualquer um pode ser um fiscal dos gastos públicos e este site oferece dados suficientes para isso.</p>
-              <p
-                class="text-justify style2"
-              >Apesar de ser especialista na fiscalização dos gastos com a verba indenizatória, a OPS também aceita denúncias de outros setores públicos, dando a elas o destino certo para que sejam apuradas pelas autoridades competentes.</p>
+
               <p class="text-justify style2">
-                O trabalho de fiscalização, coordenado por seu fundador,
+                Até o momento a OPS proporcionou
                 <a
-                  href="https://luciobig.com.br"
+                  href="https://institutoops.org.br/o-que-ja-fizemos/"
                   target="_blank"
-                >Lúcio Big</a>, é realizado voluntariamente por diversas pessoas Brasil afora.
+                  rel="nofollow"
+                  >economia os cofres públicos em valor superior a R$ 6
+                  milhões</a
+                >
+                graças a este trabalho de fiscalização que permite exigir
+                diretamente dos parlamentares a devolução dos valores
+                indevidamente utilizados.
               </p>
-            </div>
-          </div>
-        </div>
-        <div class="col-md-6">
-          <div class="card mb-3">
-            <div class="card-header bg-light">O que é CEAP?</div>
-            <div class="card-body">
+
               <p class="text-justify style2">
-                CEAP é a sigla da
-                <strong>Cota para o Exercício da Atividade Parlamentar</strong> (No Senado se chama CEAPS).
+                A OPS conta com a ajuda de cidadãos espalhados pelo Brasil e
+                exterior que ajudam diretamente nas fiscalizações que são
+                realizadas, no desenvolvimento de ferramentas de tecnologia e no
+                suporte financeiro que é feito por
+                <a
+                  href="https://institutoops.org.br/apoio/"
+                  target="_blank"
+                  rel="nofollow"
+                  >doações voluntárias</a
+                >.
               </p>
-              <p
-                class="text-justify style2"
-              >Trata-se de um recurso público disponível mensalmente aos deputados e senadores para o custeio de suas atividades parlamentares.</p>
-              <p
-                class="text-justify style2"
-              >Apesar de ser dinheiro público, fato que já implicaria no seu enquadramento à Lei Nº 8.666, o uso da verba dispensa qualquer tipo de burocracia exigida pela lei de licitações, dando ao deputado e ao senador a livre escolha da empresa a ser contratada para fornecimento do produto ou serviço.</p>
-              <p
-                class="text-justify style2"
-              >Não é raro encontrar irregularidades escandalosas, como locações de veículos feitas em padarias ou cópias de trabalhos escolares transformados em «pareceres jurídicos» generosamente pagos com o dinheiro do contribuinte.</p>
+
+              <p class="text-justify style2">
+                Apesar de ser especialista na fiscalização dos gastos com a
+                verba indenizatória, a OPS também monitora contas públicas de
+                estados e municípios, além de disponibilizar aos cidadãos um
+                <a
+                  href="https://institutoops.org.br/denunciar/"
+                  target="_blank"
+                  rel="nofollow"
+                  >canal para o recebimento de denúncias</a
+                >, anônimas ou não.
+              </p>
+
+              <p class="text-justify style2">
+                A OPS é coordenada por seu fundador,
+                <a href="https://luciobig.com.br" target="_blank" rel="nofollow"
+                  >Lúcio Big</a
+                >.
+              </p>
             </div>
           </div>
-        </div>
-         <div class="col-md-6">
           <div class="card mb-3">
-            <div class="card-header bg-light mb-3">Origem das Informações</div>
-            <ol>
-              <li class="mb-3">
-                <strong>Câmara de Deputados</strong>
-                <br />
-                <ul>
-                  <li>
-                    <a
-                      href="http://www2.camara.leg.br/transparencia/cota-para-exercicio-da-atividade-parlamentar/dados-abertos-cota-parlamentar"
-                      target="_blank"
-                      rel="nofollow"
-                    >Cota parlamentar (XML)</a>
-                  </li>
-                  <li>
-                    <a
-                      href="http://www.camara.leg.br/sitcamaraws/deputados.asmx"
-                      target="_blank"
-                      rel="nofollow"
-                    >Dados abertos (WebService)</a>
-                  </li>
-                </ul>
-              </li>
-              <li class="mb-3">
-                <strong>Senado Federal</strong>
-                <br />
-                <ul>
-                  <li>
-                    <a
-                      href="https://www12.senado.leg.br/transparencia/dados-abertos/dados-abertos-ceaps"
-                      target="_blank"
-                      rel="nofollow"
-                    >Cota parlamentar (CSV)</a>
-                  </li>
-                  <li>
-                    <a
-                      href="http://legis.senado.leg.br/dadosabertos/docs/index.html"
-                      target="_blank"
-                      rel="nofollow"
-                    >Dados abertos (WebService)</a>
-                  </li>
-                  <li>
-                    <a
-                      href="http://www.senado.gov.br/transparencia/dadosAbertos/dadosAbertos.asp"
-                      target="_blank"
-                      rel="nofollow"
-                    >Dados abertos (CSV)</a>
-                  </li>
-                </ul>
-              </li>
-              <li class="mb-3">
-                <strong>Informações dos Fornecedors</strong>
-                <ul>
-                  <li>
-                    <a
-                      href="http://www.receita.fazenda.gov.br/pessoajuridica/cnpj/cnpjreva/cnpjreva_solicitacao.asp"
-                      target="_blank"
-                      rel="nofollow"
-                    >Receita Federal</a>
-                  </li>
-                </ul>
-              </li>
-            </ol>
+            <div class="card-header bg-light">Contribua</div>
+            <div class="card-body">
+              <p>
+                <strong>Com doações</strong>
+              </p>
+              <p class="text-justify">
+                O trabalho é voluntário, mas os custos não. Ajudar a manter este
+                projeto é investir na formação da cultura do “eu fiscal”, na
+                moralização do uso de verba pública e no enfrentamento direto
+                com políticos que por vezes confundem o dinheiro público com o
+                privado. Você pode doar uma única vez ou programar doações
+                mensais.
+                <a
+                  href="https://institutoops.org.br/apoio/"
+                  target="_blank"
+                  rel="nofollow"
+                  >Clique aqui e escolha</a
+                >.
+              </p>
+            </div>
           </div>
           <div class="card mb-3">
             <div class="card-header bg-light mb-3">Informações úteis</div>
@@ -152,35 +125,41 @@
                       href="http://www2.camara.leg.br/legin/int/atomes/2009/atodamesa-43-21-maio-2009-588364-normaatualizada-cd-mesa.html"
                       target="_blank"
                       rel="nofollow"
-                    >Regras para uso da cota parlamentar (Ato da mesa Nº 43)</a>
+                      >Regras para uso da cota parlamentar (Ato da mesa Nº
+                      43)</a
+                    >
                   </li>
                   <li>
                     <a
                       href="http://www2.camara.leg.br/participe/fale-conosco/perguntas-frequentes/cota-para-o-exercicio-da-atividade-parlamentar"
                       target="_blank"
                       rel="nofollow"
-                    >Legislação relacionada</a>
+                      >Legislação relacionada</a
+                    >
                   </li>
                   <li>
                     <a
                       href="http://www.camara.gov.br/cota-parlamentar/ANEXO_ATO_DA_MESA_43_2009.pdf"
                       target="_blank"
                       rel="nofollow"
-                    >Limites mensais por parlamentar</a>
+                      >Limites mensais por parlamentar</a
+                    >
                   </li>
                   <li>
                     <a
                       href="http://www.camara.gov.br/cota-parlamentar"
                       target="_blank"
                       rel="nofollow"
-                    >Explorar a cota parlamentar na câmara</a>
+                      >Explorar a cota parlamentar na câmara</a
+                    >
                   </li>
                   <li>
                     <a
                       href="http://www2.camara.leg.br/deputados/pesquisa"
                       target="_blank"
                       rel="nofollow"
-                    >Conheça os Deputados</a>
+                      >Conheça os Deputados</a
+                    >
                   </li>
                 </ul>
               </li>
@@ -192,28 +171,32 @@
                       href="https://www12.senado.leg.br/transparencia/leg/pdf/normas/APS-5-2014-compilado.pdf"
                       target="_blank"
                       rel="nofollow"
-                    >Regras para uso da cota parlamentar</a>
+                      >Regras para uso da cota parlamentar</a
+                    >
                   </li>
                   <li>
                     <a
                       href="https://www12.senado.leg.br/transparencia/leg/legislacao-relacionada"
                       target="_blank"
                       rel="nofollow"
-                    >Legislação relacionada</a>
+                      >Legislação relacionada</a
+                    >
                   </li>
                   <li>
                     <a
                       href="http://www25.senado.leg.br/web/transparencia/sen"
                       target="_blank"
                       rel="nofollow"
-                    >Explorar a cota parlamentar no senado</a>
+                      >Explorar a cota parlamentar no senado</a
+                    >
                   </li>
                   <li>
                     <a
                       href="http://www25.senado.leg.br/web/senadores"
                       target="_blank"
                       rel="nofollow"
-                    >Conheça os senadores</a>
+                      >Conheça os senadores</a
+                    >
                   </li>
                 </ul>
               </li>
@@ -222,110 +205,202 @@
         </div>
         <div class="col-md-6">
           <div class="card mb-3">
-            <div class="card-header bg-light">Contribua</div>
+            <div class="card-header bg-light">O que é CEAP?</div>
             <div class="card-body">
-              <p>
-                <strong>Com divulgação</strong>
+              <p class="text-justify style2">
+                CEAP é a sigla da
+                <strong>Cota para o Exercício da Atividade Parlamentar</strong>
+                (No Senado se chama CEAPS).
               </p>
-              <p>Divulgue o projeto para seus amigos e contatos.</p>
-
-              <p>
-                <strong>Com análises</strong>
+              <p class="text-justify style2">
+                Trata-se de um recurso público disponível mensalmente aos
+                deputados e senadores para o custeio de suas atividades
+                parlamentares. Nos estados essa verba pode receber diversos
+                outros nomes.
               </p>
-              <p
-                class="text-justify"
-              >Você pode reportar suspeitas sobre quaisquer transações. Contribua detalhando qualquer movimentação ou padrão suspeito.</p>
-
-              <p>
-                <strong>Com código</strong>
+              <p class="text-justify style2">
+                Apesar de ser dinheiro público, fato que já implicaria no seu
+                enquadramento à Lei de Licitações, o uso da verba dispensa
+                qualquer tipo de burocracia exigida por essa lei, dando ao
+                político a livre escolha da empresa a ser contratada para
+                fornecimento do produto ou serviço. Há casos que nem mesmo
+                empresas são contratadas, apenas pessoas físicas.
               </p>
-              <p
-                class="text-justify"
-              >Esse é um projeto de código aberto, você pode contribuir com correções e novas funcionalidades.</p>
-              <p>
+              <p class="text-justify style2">
+                Não é raro encontrar irregularidades no uso desta cota. Algumas
+                são escandalosas, como locações de veículos feitas em padarias
+                ou até de pessoas que nem sabiam possuir o veículo locado pelo
+                parlamentar, até irregularidades de cunho fiscal, como a emissão
+                de notas fiscais sem lastro legal.
+              </p>
+              <p class="text-justify style2">
+                Veja alguns casos
                 <a
-                  href="https://github.com/ops-org/operacao-politica-supervisionada"
+                  href="https://institutoops.org.br/o-que-ja-fizemos/"
                   target="_blank"
-                >https://github.com/ops-org/operacao-politica-supervisionada</a>
+                  rel="nofollow"
+                  >aqui</a
+                >.
               </p>
-
-              <p>
-                <strong>Com ideias</strong>
-              </p>
-              <p
-                class="text-justify"
-              >Qualquer ideia contrutiva é sempre bem vinda. Também é possivel votar nas ideias existentes e deixar sua opinião.</p>
-              <p>
-                <a
-                  href="https://feedback.userreport.com/9cf4c59a-d438-48b0-aa5e-e16f549b9c8c/#ideas/popular"
-                  target="_blank"
-                >Deixar uma ideia</a>
-              </p>
-
-              <p>
-                <strong>Com doações</strong>
-              </p>
-              <p
-                class="text-justify"
-              >Você poderá doar uma única vez ou programar para que suas doações sejam mensais. É possível cancelar as doações recorrentes a qualquer momento, sem pagamento de taxas ou multas.</p>
-              <a
-                target="_blank"
-                href="https://sites.google.com/site/doandoparaaops/doacao-via-deposito---banco-do-brasil"
-              >
-                <img
-                  src="https://luciobigdotcomdotbr.files.wordpress.com/2015/11/bb_100.png?w=100"
-                  alt="bb"
-                  width="96"
-                  height="56"
-                />
-              </a>
-              <a target="_blank" href="https://www.patreon.com/bePatron?patAmt=20&amp;u=192467">
-                <img
-                  src="https://luciobigdotcomdotbr.files.wordpress.com/2015/11/patreon_25.png?w=100"
-                  alt="patreon"
-                  width="96"
-                  height="56"
-                />
-              </a>
-              <a
-                target="_blank"
-                href="https://sites.google.com/site/doandoparaaops/doacao-mensal-paypal"
-              >
-                <img
-                  src="https://luciobigdotcomdotbr.files.wordpress.com/2015/11/paypal_25.png?w=100"
-                  alt="paypal"
-                  width="96"
-                  height="56"
-                />
-              </a>
-              <a
-                target="_blank"
-                href="https://pagseguro.uol.com.br/checkout/nc/nl/donation/sender-identification.jhtml?t=b71889bbdc400e71008fab3d4ba4802662561764f8d69f10#rmcl"
-              >
-                <img
-                  src="https://luciobigdotcomdotbr.files.wordpress.com/2015/11/pagseguro_25.png?w=100"
-                  alt="pagseguro"
-                  width="96"
-                  height="56"
-                />
-              </a>
-              <a target="_blank" href="https://apoia.se/ops">
-                <img
-                  src="https://luciobigdotcomdotbr.files.wordpress.com/2015/11/apo.png?w=100"
-                  alt="apoia.se"
-                  width="96"
-                  height="56"
-                />
-              </a>
             </div>
+          </div>
+          <div class="card mb-3">
+            <div class="card-header bg-light mb-3">Origem das Informações</div>
+            <ol>
+              <li class="mb-3">
+                <strong>Câmara de Deputados</strong>
+                <br />
+                <ul>
+                  <li>
+                    <a
+                      href="https://www.camara.leg.br/transparencia/gastos-parlamentares"
+                      target="_blank"
+                      rel="nofollow"
+                      >Cota parlamentar</a
+                    >
+                  </li>
+                </ul>
+                <ul>
+                  <li>
+                    <a
+                      href="https://dadosabertos.camara.leg.br/"
+                      target="_blank"
+                      rel="nofollow"
+                      >Dados Abertos</a
+                    >
+                  </li>
+                </ul>
+                <ul>
+                  <li>
+                    <a
+                      href="https://www.camara.leg.br/deputados/quem-sao"
+                      target="_blank"
+                      rel="nofollow"
+                      >Conheça os Deputados</a
+                    >
+                  </li>
+                </ul>
+              </li>
+              <li class="mb-3">
+                <strong>Senado Federal</strong>
+                <br />
+                <ul>
+                  <li>
+                    <a
+                      href="https://www12.senado.leg.br/dados-abertos/conjuntos?portal=Administrativo&grupo=senadores"
+                      target="_blank"
+                      rel="nofollow"
+                      >Cota parlamentar (CSV)</a
+                    >
+                  </li>
+                </ul>
+                <ul>
+                  <li>
+                    <a
+                      href="https://www25.senado.leg.br/web/senadores/em-exercicio"
+                      target="_blank"
+                      rel="nofollow"
+                      >Conheça os Senadores</a
+                    >
+                  </li>
+                </ul>
+              </li>
+
+              <li class="mb-3">
+                <strong>Assembleias Legislativas</strong>
+                <ul>
+                  <li>
+                    <a
+                      href="https://www.al.ba.gov.br/transparencia/prestacao-contas"
+                      target="_blank"
+                      rel="nofollow"
+                      >Bahia</a
+                    >
+                  </li>
+                  <li>
+                    <a
+                      href="https://dadosabertos.cl.df.gov.br/View/verbas.html"
+                      target="_blank"
+                      rel="nofollow"
+                      >Distrito Federal</a
+                    >
+                  </li>
+                  <li>
+                    <a
+                      href="https://transparencia.al.go.leg.br/gestao-parlamentar/verba-indenizatoria"
+                      target="_blank"
+                      rel="nofollow"
+                      >Goiás</a
+                    >
+                  </li>
+                  <li>
+                    <a
+                      href="https://www.almg.gov.br/acompanhe/prestacao_contas/index.html?aba=js_tabVerba"
+                      target="_blank"
+                      rel="nofollow"
+                      >Minas Gerais</a
+                    >
+                  </li>
+                  <li>
+                    <a
+                      href="https://consultas.assembleia.pr.leg.br/#/ressarcimento"
+                      target="_blank"
+                      rel="nofollow"
+                      >Paraná</a
+                    >
+                  </li>
+                  <li>
+                    <a
+                      href="https://transparencia.alesc.sc.gov.br/gabinetes.php"
+                      target="_blank"
+                      rel="nofollow"
+                      >Santa Catarina</a
+                    >
+                  </li>
+                  <li>
+                    <a
+                      href="https://www.al.sp.gov.br/dados-abertos/recurso/21"
+                      target="_blank"
+                      rel="nofollow"
+                      >São Paulo</a
+                    >
+                  </li>
+                </ul>
+              </li>
+
+              <li class="mb-3">
+                <strong>Fornecedores</strong>
+                <ul>
+                  <li>
+                    <a
+                      href="http://www.receita.fazenda.gov.br/pessoajuridica/cnpj/cnpjreva/cnpjreva_solicitacao.asp"
+                      target="_blank"
+                      rel="nofollow"
+                      >Receita Federal</a
+                    >
+                    (via
+                    <a
+                      href="https://minhareceita.org"
+                      target="_blank"
+                      rel="nofollow"
+                      >Minha Receita</a
+                    >)
+                  </li>
+                </ul>
+              </li>
+            </ol>
           </div>
         </div>
       </div>
     </div>
     <br />
-    <div style="padding-bottom: 10px;">
+    <div style="padding-bottom: 10px">
       <p class="text-right">
-        <a href="https://luciobig.com.br/o-que-e-ops/" target="_blank" class="btn btn-light btn-sm">
+        <a
+          href="https://luciobig.com.br/o-que-e-ops/"
+          target="_blank"
+          class="btn btn-light btn-sm"
+        >
           Saiba mais sobre a OPS&nbsp;
           <i class="fa fa-chevron-right"></i>
         </a>
